Add files via upload
ROS-SpiNNaker proof of concept of ROS and SpiNNaker-5 (Spiking Neural Network) working in tandem where ROS generates random numbers and there are 100 injecting neurons and 10 spiking neurons in (population called pop_forward) in the population. As a message (integer number) comes from ROS the injecting neuron with that number as neuron_id is spiked and connectivity elicits a spike in the pop_forward population. Once a spike is generated in this population, the neuron_id of neuron which spiked is sent back to the ROS node.